+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New features:
 Bug fixes:
 
 * Fix `Marshal.load` of multiple `Symbols` with an explicit encoding (#1624).
+* Fix `rb_str_modify_expand` to preserve existing bytes (#2392).
 
 Compatibility: